requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1
psycopg2-binary>=2.9.9

# Caching
redis>=5.0.1
hiredis>=2.3.2

# OpenGradient
opengradient>=0.2.1

# Bots
aiogram>=3.4.1
discord.py>=2.3.2

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.1

# Async Support
nest-asyncio>=1.5.8

# Validation
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Logging
loguru>=0.7.2

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0

# Utilities
tenacity>=8.2.3
